engine/requirements.txt
transformers
diffusers
accelerate
safetensors
gguf
bitsandbytes
matplotlib
pyttsx3
chattts
kokoro
moviepy
opencv-python
scikit-learn
scikit-image
qwen-vl-utils
sentencepiece
peft
av
pydub
soundfile
librosa
pysptk
pyworld
torch
torchvision
torchaudio
speechbrain
webrtcvad
gfpgan
munch
tqdm
numba
yacs
kornia
facexlib
asteval
arxiv
wikipedia-api
firecrawl-py
PyMuPDF
html-to-markdown[lxml]